Homotopies

Homotopies

Pre-defined homotopies

HomotopyContinuation.Homotopies.StraightLineHomotopy — Type.

StraightLineHomotopy(G, F; gamma=exp(i * 2π*rand()))

Construct the homotopy $H(x, t) = γtG(x) + (1-t)F(x)$.

source

HomotopyContinuation.Homotopies.FixedPointHomotopy — Type.

FixedPointHomotopy(F, x₀; gamma=exp(i * 2π*rand()))

Construct the homotopy $H(x, t) = (1-t)F(x) + γt(x-x₀)$.

source

Interface for custom homotopies

Abstract types

HomotopyContinuation.HomotopiesBase.AbstractHomotopy — Type.

AbstractHomotopy

Representing a homotopy.

source

HomotopyContinuation.HomotopiesBase.AbstractHomotopyCache — Type.

AbstractHomotopyCache

A cache to avoid allocations for the evaluation of an AbstractHomotopy.

source

Mandatory

HomotopyContinuation.HomotopiesBase.evaluate! — Function.

evaluate!(u, H::AbstractHomotopy, x, t, cache::AbstractHomotopyCache)

Evaluate the homotopy H at (x, t) and store the result in u.

source

HomotopyContinuation.HomotopiesBase.jacobian! — Function.

jacobian!(u, H::AbstractHomotopy, x, t, cache::AbstractHomotopyCache)

Evaluate the Jacobian of the homotopy H at (x, t) and store the result in u.

source

HomotopyContinuation.HomotopiesBase.dt! — Function.

dt!(u, H::AbstractHomotopy, x, t, cache::AbstractHomotopyCache)

Evaluate the homotopy H at (x, t) and store the result in u.

source

Base.size — Method.

Base.size(H::AbstractHomotopy)

Returns a tuple (m, n) indicating that H is a homotopy of m polynomials m in n variables.

source

Optional

HomotopyContinuation.HomotopiesBase.cache — Function.

cache(H::AbstractHomotopy, x, t)

Create a cache for the evaluation (incl. Jacobian) of F with elements of the type of x.

source

HomotopyContinuation.HomotopiesBase.evaluate_and_jacobian! — Function.

evaluate_and_jacobian!(u, U, F, x, t, cache::AbstractHomotopyCache)

Evaluate the homotopy H and its Jacobian at (x, t) and store the results in u (evalution) and U (Jacobian).

source

HomotopyContinuation.HomotopiesBase.evaluate_and_jacobian — Function.

evaluate_and_jacobian(H::AbstractHomotopy, x, t, cache::AbstractHomotopyCache)

Evaluate the homotopy H and its Jacobian at (x, t).

source

HomotopyContinuation.HomotopiesBase.jacobian_and_dt! — Function.

jacobian_and_dt!(U, u, H, x, t, cache::AbstractHomotopyCache)

Evaluate the homotopy H and its derivative w.r.t. t at (x, t) and store the results in u (evalution) and v (∂t).

source

HomotopyContinuation.HomotopiesBase.evaluate — Function.

evaluate(H::AbstractHomotopy, x, t, cache::AbstractHomotopyCache)

Evaluate the homotopy H at (x, t).

source

HomotopyContinuation.HomotopiesBase.jacobian — Function.

jacobian(H::AbstractHomotopy, x, t, cache::AbstractHomotopyCache)

Evaluate the Jacobian of the homotopy H at (x, t).

source

HomotopyContinuation.HomotopiesBase.dt — Function.

dt(H::AbstractHomotopy, x::AbstractVector, cache::AbstractHomotopyCache)

Evaluate the homotopy H at (x, t).

source

HomotopyContinuation.HomotopiesBase.precondition! — Function.

precondition!(H::AbstractHomotopy, x, t, cache)

Prepare a homotopy for things like pathtracking starting at x and t. This can modify x as well as H and anything in cache. By default this is a no-op. If H wraps another homotopy this should call precondition! on this as well.

source

HomotopyContinuation.HomotopiesBase.update! — Function.

update!(H::AbstractHomotopy, x, t, cache)

Update a homotopy for new values of x and x, i.e., update an affine patch. This can modify x as well as H and anything in cache. By default this is a no-op. If H wraps another homotopy this should call update! on this as well.

source