Predictors and Correctors

Predictors and Correctors

We use a predictor-corrector scheme to track paths. These are the predictors and correctors currently available.

Predictors

The following predictors are currently implemented.

HomotopyContinuation.Predictors.RK4 — Type.

RK4()

The classical Runge-Kutta predictor of order 4.

source

HomotopyContinuation.Predictors.Euler — Type.

Euler()

This uses the explicit Euler method for prediction, also known as the tangent predictor.

source

HomotopyContinuation.Predictors.NullPredictor — Type.

NullPredictor()

A predictor which does no prediction step, i.e., it just returns the input as its prediction.

source

Correctors

HomotopyContinuation.Correctors.Newton — Type.

Newton()

A classical simple Newton operator for square linear systems using the LU factorization to solve the linear systems.

source