Reference

Reference

Input

We support any polynomials which follow the MultivariatePolynomials interface. By default we export the routines @polyvar, PolyVar, differentiate and variables from the DynamicPolynomials implementation. With these you can simply create variables

# Create variables x, y, z
@polyvar x y z
f = x^2+y^2+z^2

# You can also create an array of variables
@polyvar x[1:3] # This creates x1, x2, x3 accessed by x[1], x[2], x[3]
f = dot(x, x) # = x[1]^2+x[2]^2+x[3]^2

# Also you can create matrices of variables
# This creates x1_1, x1_2, x2_1, x2_2 accessed by
# x[1,1], x[1,2], x[2,1], x[2,2]
@polyvar x[1:2, 1:2]

AffinePatches

Affine patches are there to augment projective system such that they can be considered as (locally) affine system. By default the following patches are defined

HomotopyContinuation.AffinePatches.OrthogonalPatch — Type.

OrthogonalPatch()

source

HomotopyContinuation.AffinePatches.EmbeddingPatch — Type.

EmbeddingPatch()

Holds an AbstractProjectiveVector onto its affine patch. With this the effect is basically the same as tracking in affine space.

source

HomotopyContinuation.AffinePatches.RandomPatch — Type.

RandomPatch()

A random patch. The vector has norm 1.

source

HomotopyContinuation.AffinePatches.FixedPatch — Type.

FixedPatch()

source

Interface

Each patch has to follow the following interface.

HomotopyContinuation.AffinePatches.AbstractAffinePatch — Type.

AbstractAffinePatch

An affine patch is a hyperplane defined by $v⋅x-1=0$.

source

HomotopyContinuation.AffinePatches.state — Function.

state(::AbstractAffinePatch, x)::AbstractAffinePatchState

Construct the state of the path from x.

source

HomotopyContinuation.AffinePatches.precondition! — Function.

precondition!(v::AbstractAffinePatchState, x)

Modify both such that v is properly setup and v⋅x-1=0 holds.

source

HomotopyContinuation.AffinePatches.update! — Function.

update_patch!(::AbstractAffinePatchState, x)

Update the patch depending on the local state.

source