Solving

Solving systems

solve

HomotopyContinuation.solve — Function.

solve(F; options...)

Solve the system F using a total degree homotopy. F can be

Vector{<:MultivariatePolynomials.AbstractPolynomial} (e.g. constructed by @polyvar)

Systems.AbstractSystem (the system has to represent a homogenous polynomial system.)

Example

Assume we want to solve the system $F(x,y) = (x^2+y^2+1, 2x+3y-1)$.

@polyvar x y
solve([x^2+y^2+1, 2x+3y-1])

If you polynomial system is already homogenous, but you would like to consider it as an affine system you can do

@polyvar x y z
solve([x^2+y^2+z^2, 2x+3y-z], homvar=z)

This would result in the same result as solve([x^2+y^2+1, 2x+3y-1]).

To solve $F$ by a custom Systems.AbstractSystem you can do

@polyvar x y z
# The system `F` has to be homgoenous system
F = Systems.SPSystem([x^2+y^2+z^2, 2x+3y-z]) # Systems.SPSystem <: Systems.AbstractSystem
# To solve the original affine system we have to tell that the homogenization variable has index 3
solve(F, homvar=3)

Start Target Homotopy

solve(G, F, start_solutions; options...)

Solve the system F by tracking the each provided solution of G (as provided by start_solutions).

Example

@polyvar x y
G = [x^2-1,y-1]
F = [x^2+y^2+z^2, 2x+3y-z]
solve(G, F, [[1, 1], [-1, 1]])

Parameter Homotopy

solve(F::Vector{<:MultivariatePolynomials.AbstractPolynomial}, parametervariables, startparameters, targetparameters, startsolutions)

Construct a parameter homotopy $H(x,t) := F(x; tp₁+(1-t)p₀)$ where p₁ is startparameters, p₀ is targetparameters and the parametervariables are the variables of F which should be considerd parameters.

Example

We want to solve a parameter homotopy $H(x,t) := F(x; t[1, 0]+(1-t)[2, 4])$ where

\[F(x; a) := (x₁^2-a₁, x₁x₂-a₁+a₂)\]

and let's say we are only intersted in tracking of $[1,1]$. This can be accomplished as follows

@polyvar x[1:2] a[1:2]
F = [x[1]^2-a[1], x[1]*x[2]-a[1]+a[2]]
p₁ = [1, 0]
p₀ = [2, 4]
startsolutions = [[1, 1]]
solve(F, a, p₁, p₀, startsolutions)

Abstract Homotopy

solve(H::Homotopies.AbstractHomotopy, start_solutions; options...)

Solve the homotopy H by tracking the each solution of $H(⋅, t)$ (as provided by start_solutions) from $t=1$ to $t=0$.

Options

General options:

seed::Int: The random seed used during the computations.

homvar::Union{Int,MultivariatePolynomials.AbstractVariable}: This considers the homogenous system F as an affine system which was homogenized by homvar. If F is an AbstractSystem homvar is the index (i.e. Int) of the homogenization variable. If F is an AbstractVariables (e.g. created by @polyvar x) homvar is the actual variable used in the system F.

endgame_start=0.1: The value oft` for which the endgame is started.

report_progress=true: Whether a progress bar should be printed to STDOUT.

Pathtracking specific:

corrector::Correctors.AbstractCorrector: The corrector used during in the predictor-corrector scheme. The default is Correctors.Newton.

corrector_maxiters=2: The maximal number of correction steps in a single step.

predictor::Predictors.AbstractPredictor: The predictor used during in the predictor-corrector scheme. The default is Predictors.RK4.

refinement_maxiters=corrector_maxiters: The maximal number of correction steps used to refine the final value.

refinement_tol=1e-11: The precision used to refine the final value.

tol=1e-7: The precision used to track a value.

initial_steplength=0.1: The initial step size for the predictor.

steplength_increase_factor=2.0: The factor with which the step size is increased after steplength_consecutive_successes_necessary consecutive successes.

steplength_decrease_factor=inv(increase_factor): The factor with which the step size is decreased after a step failed.

steplength_consecutive_successes_necessary=5: The numer of consecutive successes necessary until the step size is increased by steplength_increase_factor.

maximal_steplength=max(0.1, initial_steplength): The maximal step length.

minimal_steplength=1e-14: The minimal step size. If the size of step is below this the path is considered failed.

Endgame specific options

cauchy_loop_closed_tolerance=1e-3: The tolerance for which is used to determine whether a loop is closed. The distance between endpoints is normalized by the maximal difference between any point in the loop and the starting point.

cauchy_samples_per_loop=6: The number of samples used to predict an endpoint. A higher number of samples should result in a better approximation. Note that the error should be roughly $t^n$ where $t$ is the current time of the loop and $n$ is cauchy_samples_per_loop.

egtol=1e-10: This is the tolerance necessary to declare the endgame converged.

maxnorm=1e5: If our original problem is affine we declare a path at infinity if the infinity norm with respect to the standard patch is larger than maxnorm.

maxwindingnumber=15: The maximal windingnumber we try to find using Cauchys integral formula.

max_extrapolation_samples=4: During the endgame a Richardson extrapolation is used to improve the accuracy of certain approximations. This is the maximal number of samples used for this.

minradius=1e-15: A path is declared false if the endgame didn't finished until then.

sampling_factor=0.5: During the endgame we approach $0$ by the geometric series $h^kR₀$ where $h$ is sampling_factor and R₀ the endgame start provided in runendgame.

source

Depending on the input solve returns one of the following types

HomotopyContinuation.Solving.AffineResult — Type.

AffineResult <: Result

The result of an (non-homogenous) system of polynomials.

source

HomotopyContinuation.Solving.ProjectiveResult — Type.

ProjectiveResult <: Result

The result of a homogenous system of polynomials.

source

Result

HomotopyContinuation.Solving.finite — Function.

finite(result::AffineResult)

Return all PathResults for which the result is successfull and the contained solution is indeed a solution of the system.

finite(f::Function, result)

Additionally you can apply a transformation f on each result.

source

HomotopyContinuation.Solving.results — Function.

results(result; onlyreal=false, realtol=1e-6, onlynonsingular=false, singulartol=1e10, onlyfinite=true)

Return all PathResults for which the given conditions apply.

results(f::Function, result; kwargs...)

Additionally you can apply a transformation f on each result.

Example

R = solve(F)

# This gives us all solutions considered real (but still as a complex vector).
realsolutions = results(solution, R, onlyreal=true)

source

HomotopyContinuation.Solving.solutions — Function.

solutions(result, onlyreal=Val{false};realtol=1e-6, onlynonsingular=false, singulartol=1e10, onlyfinite=true)

Return all solution (as Vectors) for which the given conditions apply. If onlyreal is Val{true} this returns the real solutions. This is different than applying results(solution, result, onlyreal=true).

Example

```julia julia> @polyvar x y julia> result = solve([(x-2)y, y+x+3]); julia> solutions(result, Val{true}) [[2.0, -5.0], [-3.0, 0.0]] julia> solutions(result) [[2.0+0.0im, -5.0+0.0im], [-3.0+0.0im, 0.0+0.0im]]

source

HomotopyContinuation.Solving.failed — Function.

failed(result)

Get all results where the path tracking failed.

source

HomotopyContinuation.Solving.atinfinity — Function.

atinfinity(result::AffineResult)

Get all results where the solutions is at infinity.

source

HomotopyContinuation.Solving.singular — Function.

singular(result; tol=1e10)

Get all singular solutions. A solution is considered singular if its windingnumber is larger than 1 or the condition number is larger than tol.

source

HomotopyContinuation.Solving.nonsingular — Function.

nonsingular(result::AffineResult)

Return all PathResults for which the solution is non-singular.

source

HomotopyContinuation.Solving.seed — Function.

seed(result)

The random seed used in the computation.

source

HomotopyContinuation.Solving.nresults — Function.

nresults(result)

The number of proper solutions.

source

HomotopyContinuation.Solving.nfinite — Function.

nresults(affineresult)

The number of finite solutions.

source

HomotopyContinuation.Solving.nsingular — Function.

nsingular(result; tol=1e10)

The number of singular solutions. A solution is considered singular if its windingnumber is larger than 1 or the condition number is larger than tol.

source

HomotopyContinuation.Solving.nnonsingular — Function.

nnonsingular(result; tol=1e-10)

The number of non-singular solutions.

source

HomotopyContinuation.Solving.natinfinity — Function.

natinfinity(affineresult)

The number of solutions at infinity.

source

HomotopyContinuation.Solving.nfailed — Function.

nafailed(result)

The number of failed paths.

source

PathResult

HomotopyContinuation.Solving.solution — Function.

solution(pathresult)

Get the solution of the path.

source

HomotopyContinuation.Solving.residual — Function.

residual(pathresult)

Get the residual of the solution $x$ of the path, i.e., $||H(x, t)||_{\infty}$.

source

HomotopyContinuation.Solving.start_solution — Function.

residual(pathresult)

Get the residual of the solution $x$ of the path, i.e., $||H(x, t)||_{\infty}$.

source

HomotopyContinuation.Solving.issuccess — Function.

issuccess(pathresult)

Checks whether the path is successfull.

source

HomotopyContinuation.Solving.isfailed — Function.

isfailed(pathresult)

Checks whether the path failed.

source

HomotopyContinuation.Solving.isaffine — Function.

isaffine(pathresult)

Checks whether the path result is affine.

source

HomotopyContinuation.Solving.isprojective — Function.

isprojective(pathresult)

Checks whether the path result is projective.

source

HomotopyContinuation.Solving.isatinfinity — Function.

isatinfinity(pathresult)

Checks whether the path goes to infinity.

source

HomotopyContinuation.Solving.issingular — Function.

issingular(pathresult; tol=1e10)
issingular(pathresult, tol)

Checks whether the path result is singular. This is true if the winding number > 1 or if the condition number of the Jacobian is larger than tol.

source

HomotopyContinuation.Solving.isnonsingular — Function.

isnonsingular(pathresult; tol=1e10)

Checks whether the path result is non-singular. This is true if it is not singular.

source