Systems

Polynomial systems

Pre-defined systems

HomotopyContinuation.Systems.FPSystem — Type.

FPSystem(polynomials, vars) <: AbstractSystem

Create a system using the FixedPolynomials package.

source

HomotopyContinuation.Systems.SPSystem — Type.

SPSystem(polynomials, vars) <: AbstractSystem

Create a system using the StaticPolynomials package.

source

Interface for custom systems

Abstract types

HomotopyContinuation.SystemsBase.AbstractSystem — Type.

AbstractSystem

Representing a system of polynomials.

source

HomotopyContinuation.SystemsBase.AbstractSystemCache — Type.

AbstractSystemCache

A cache to avoid allocations for the evaluation of an AbstractSystem.

source

Mandatory

HomotopyContinuation.SystemsBase.evaluate! — Function.

evaluate!(u, F::AbstractSystem, x [, cache::AbstractSystemCache])

Evaluate the system F at x and store the result in u.

source

HomotopyContinuation.SystemsBase.evaluate — Function.

evaluate(F::AbstractSystem, x::AbstractVector [, cache::AbstractSystemCache])

Evaluate the system F at x.

source

HomotopyContinuation.SystemsBase.jacobian! — Function.

jacobian!(u, F::AbstractSystem, x [, cache::AbstractSystemCache])

Evaluate the Jacobian of the system F at x and store the result in u.

source

HomotopyContinuation.SystemsBase.jacobian — Function.

jacobian(F::AbstractSystem, x [, cache::AbstractSystemCache])

Evaluate the Jacobian of the system F at x.

source

Base.size — Method.

Base.size(F::AbstractSystem)

Returns a tuple (m, n) indicating that F is a system of m polynomials m in n variables.

source

Optional

HomotopyContinuation.SystemsBase.cache — Function.

cache(F::AbstractSystem, x)

Create a cache for the evaluation (incl. Jacobian) of F with elements of the type of x. The default implementation returns a NullCache.

source

HomotopyContinuation.SystemsBase.evaluate_and_jacobian! — Function.

evaluate_and_jacobian!(u, U, F, x [, cache::AbstractSystemCache])

Evaluate the system F and its Jacobian at x and store the results in u (evalution) and U (Jacobian).

source

HomotopyContinuation.SystemsBase.evaluate_and_jacobian — Function.

evaluate_and_jacobian(F::AbstractSystem, x [, cache::AbstractSystemCache])

Evaluate the system F and its Jacobian at x.

source