Systems and Homotopies

Systems (AbstractSystem) and homotopies (AbstractHomotopy) are used for the numerical computations.

Systems

Systems (AbstractSystem) are the basic building blocks of homotopies.

HomotopyContinuation.AbstractSystem — Type

AbstractSystem

An abstract type representing a polynomial system $F(x)$.

The following systems are available:

AffineChartSystem

CompositionSystem

FixedParameterSystem

ModelKitSystem

RandomizedSystem

AffineChartSystem

HomotopyContinuation.AffineChartSystem — Type

AffineChartSystem(F::AbstractSystem, v::PVector{T,N})

Given a system $F(x): (ℙ^{m_1} × ⋯ × ℙ^{m_N}) → ℂⁿ$ this creates a new affine system $F′$ which operates on the affine chart defined by the vector $v ∈ ℙ^{m_1} × ⋯ × ℙ^{m_N}$ and the augmented conditions $vᵀx = 1$.

HomotopyContinuation.on_affine_chart — Method

on_affine_chart(F::Union{System,AbstractSystem}, dimensions)

Construct an AffineChartSystem on a randomly generated chart v. Each entry is drawn idepdently from a univariate normal distribution.

CompositionSystem

HomotopyContinuation.CompositionSystem — Type

CompositionSystem(G::AbstractSystem, F::AbstractSystem)

Construct the system $G(F(x;p);p)$. Note that the parameters are passed to $G$ and $F$ are identical.

HomotopyContinuation.compose — Function

compose(G::Union{AbstractSystem,System}, F::Union{AbstractSystem,System})

Construct the composition $G(F(x))$. You can also use the infix operator ∘ (written by \circ).

Example

julia> @var a b c x y z

julia> g = System([a * b * c]);

julia> f = System([x+y, y + z, x + z]);

julia> compose(g, f)
Composition G ∘ F:
F:
ModelKitSystem{(0xbb16b481c0808501, 1)}:
Compiled: System of length 3
 3 variables: x, y, z

 x + y
 y + z
 x + z

G:
ModelKitSystem{(0xf0a2384a42428501, 1)}:
Compiled: System of length 1
 3 variables: a, b, c

 a*b*c

julia> (g ∘ f)([x,y,z])
1-element Array{Expression,1}:
 (x + z)*(y + z)*(x + y)

FixedParameterSystem

HomotopyContinuation.FixedParameterSystem — Type

FixedParameterSystem(F:AbstractSystem, parameters)

Construct a system from the given AbstractSystem F with the given parameters fixed.

HomotopyContinuation.fix_parameters — Function

fix_parameters(F::AbstractSystem, p)

Fix the parameters of the given system F. Returns a FixedParameterSystem.

ModelKitSystem

HomotopyContinuation.ModelKitSystem — Type

ModelKitSystem(F:System)

Construct a system from the given System F. The difference to F is that this compiles a straight line programm for the fast evaluation of F and that ModelKitSystem <: AbstractSystem.

RandomizedSystem

HomotopyContinuation.RandomizedSystem — Type

RandomizedSystem(F::Union{System,AbstractSystem}, k::Integer) <: AbstractSystem

Given a $n × N$ system F with $n > N$ this constructs the system $\mathfrak{R}(F; k)(x) = [I A]⋅F(x)$ where $I$ is a $k × k$ identity matrix and $A$ is random complex $k × n$ matrix. See Chapter 13.5 in [SW05] for more details.

RandomizedSystem(F::Union{System,AbstractSystem}, A::Matrix{ComplexF64})

Explicitly provide the used randomization matrix A.

Homotopies

Homotopies (AbstractHomotopy) are at the heart of homotopy continuation. A homotopy is a parameterized family $H(x,t)$ of polynomial systems. By convention, homotopies are tracked from $t=1$ to $t=0$, i.e., $H(x,1)$ is considered the start system and $H(x,0)$ is the target system.

HomotopyContinuation.AbstractHomotopy — Type

AbstractHomotopy

An abstract type representing a homotopy $H(x,t)$.

The following homotopies are available:

AffineChartHomotopy

AffineSubspaceHomotopy

ModelKitHomotopy

ParameterHomotopy

StraightLineHomotopy

AffineChartHomotopy

HomotopyContinuation.AffineChartHomotopy — Type

AffineChartHomotopy(H::AbstractHomotopy, v::PVector{T,N})

Given a homotopy $H(x,t): (ℙ^{m_1} × ⋯ × ℙ^{m_N}) × ℂ → ℂⁿ$ this creates a new affine homotopy $H̄$ which operates on the affine chart defined by the vector $v ∈ ℙ^{m_1} × ⋯ × ℙ^{m_N}$ and the augmented conditions $vᵀx = 1$.

HomotopyContinuation.on_affine_chart — Method

on_affine_chart(H::Union{Homotopy,AbstractHomotopy}, proj_dims)

Construct an AffineChartHomotopy on a randomly generated chart v. Each entry is drawn idepdently from a univariate normal distribution.

AffineSubspaceHomotopy

HomotopyContinuation.AffineSubspaceHomotopy — Type

AffineSubspaceHomotopy(F::System, V::AffineSubspace, W::AffineSubspace)
AffineSubspaceHomotopy(F::AbstractSystem, V::AffineSubspace, W::AffineSubspace)

Creates a homotopy $H(x,t) = (F ∩ γ(t))(x)$ where $γ(t)$ is a family of affine subspaces such that $H(x,1) = (F ∩ V)(x)$ and $H(x,0) = (F ∩ W)(x)$. Here $γ(t)$ is the geodesic between V and W in the affine Grassmanian, i.e., it is the curve of minimal length connecting V and W. See also AffineSubspace and geodesic and the references therein.

HomotopyContinuation.set_subspaces! — Function

set_subspaces!(H::AffineSubspaceHomotopy, start::AffineSubspace, target::AffineSubspace)

Update the homotopy H to track from the affine subspace start to target.

ModelKitHomotopy

HomotopyContinuation.ModelKitHomotopy — Type

ModelKitHomotopy(H::Homotopy, parameters = nothing)

Construct a homotopy from the given homotopy H with the given parameters fixed. The difference to H is that this compiles a straight line programm for the fast evaluation of H and that ModelKitHomotopy <: AbstractHomotopy.

ParameterHomotopy

HomotopyContinuation.ParameterHomotopy — Type

ParameterHomotopy(F::Union{AbstractSystem,System}; start_parameters, target_parameters)
ParameterHomotopy(F::Union{AbstractSystem,System}, start_parameters, target_parameters)

Construct the parameter homotopy $H(x,t) = F(x; t p + (1 - t) q)$ where $p$ is start_parameters and $q$ is target_parameters.

StraightLineHomotopy

HomotopyContinuation.StraightLineHomotopy — Type

StraightLineHomotopy(G::System, F::System; gamma = 1.0)
StraightLineHomotopy(G::AbstractSystem, F::AbstractSystem; gamma = 1.0)

Constructs the straight line homotopy $H(x, t) = γ t G(x) + (1-t) F(x)$ where $γ$ is gamma.

SW05Sommese, A. J., & Wampler, C. W. (2005). The Numerical Solution of Systems of Polynomials Arising in Engineering and Science. World Scientific.