Solving

Solving systems

solve

HomotopyContinuation.solve — Function.

solve(F::Vector{<:MP.AbstractPolynomial}; options...)

Solve the system F using a total degree homotopy.

solve(G, F, start_solutions; options...)

Solve the system F by tracking the each solution of G (as provided by start_solutions).

source

Depending on the input solve returns one of the following types

HomotopyContinuation.Solving.AffineResult — Type.

AffineResult <: Result

The result of an (non-homogenous) system of polynomials.

source

HomotopyContinuation.Solving.ProjectiveResult — Type.

ProjectiveResult <: Result

The result of a homogenous system of polynomials.

source

Result

HomotopyContinuation.Solving.finite — Function.

finite(result::AffineResult)

Return all PathResults for which the result is successfull and the contained solution is indeed a solution of the system.

finite(f::Function, result)

Additionally you can apply a transformation f on each result.

source

HomotopyContinuation.Solving.results — Function.

results(result; only_real=false, realtol=1e-6, onlynonsingular=true, singulartol=1e10, onlyfinite=true)

Return all PathResults for which the given conditions apply.

results(f::Function, result; kwargs...)

Additionally you can apply a transformation f on each result.

Example

```julia R = solve(F)

This gives us all solutions considered real (but still as a complex vector).

realsolutions = results(solution, R, only_real=true)

source

HomotopyContinuation.Solving.failed — Function.

failed(result)

Get all results where the path tracking failed.

source

HomotopyContinuation.Solving.atinfinity — Function.

atinfinity(result::AffineResult)

Get all results where the solutions is at infinity.

source

HomotopyContinuation.Solving.singular — Function.

singular(result; tol=1e10)

Get all singular solutions. A solution is considered singular if its windingnumber is larger than 1 or the condition number is larger than tol.

source

HomotopyContinuation.Solving.nonsingular — Function.

nonsingular(result::AffineResult)

Return all PathResults for which the solution is non-singular.

source

HomotopyContinuation.Solving.seed — Function.

seed(result)

The random seed used in the computation.

source

HomotopyContinuation.Solving.nresults — Function.

nresults(result)

The number of proper solutions.

source

HomotopyContinuation.Solving.nfinite — Function.

nresults(affineresult)

The number of finite solutions.

source

HomotopyContinuation.Solving.nsingular — Function.

nsingular(result; tol=1e10)

The number of singular solutions. A solution is considered singular if its windingnumber is larger than 1 or the condition number is larger than tol.

source

HomotopyContinuation.Solving.nnonsingular — Function.

nnonsingular(result)

The number of non-singular solutions.

source

HomotopyContinuation.Solving.natinfinity — Function.

natinfinity(affineresult)

The number of solutions at infinity.

source

HomotopyContinuation.Solving.nfailed — Function.

nafailed(result)

The number of failed paths.

source

PathResult

HomotopyContinuation.Solving.solution — Function.

solution(pathresult)

Get the solution of the path.

source

HomotopyContinuation.Solving.residual — Function.

residual(pathresult)

Get the residual of the solution $x$ of the path, i.e., $||H(x, t)||_{\infty}$.

source

HomotopyContinuation.Solving.start_solution — Function.

residual(pathresult)

Get the residual of the solution $x$ of the path, i.e., $||H(x, t)||_{\infty}$.

source

HomotopyContinuation.Solving.issuccess — Function.

issuccess(pathresult)

Checks whether the path is successfull.

source

HomotopyContinuation.Solving.isfailed — Function.

isfailed(pathresult)

Checks whether the path failed.

source

HomotopyContinuation.Solving.isaffine — Function.

isaffine(pathresult)

Checks whether the path result is affine.

source

HomotopyContinuation.Solving.isprojective — Function.

isprojective(pathresult)

Checks whether the path result is projective.

source

HomotopyContinuation.Solving.isatinfinity — Function.

isatinfinity(pathresult)

Checks whether the path goes to infinity.

source

HomotopyContinuation.Solving.issingular — Function.

issingular(pathresult; tol=1e10)
issingular(pathresult, tol)

Checks whether the path result is singular. This is true if the winding number > 1 or if the condition number of the Jacobian is larger than tol.

source

HomotopyContinuation.Solving.isnonsingular — Function.

isnonsingular(pathresult; tol=1e10)

Checks whether the path result is non-singular. This is true if it is not singular.

source